Update the readme and sort out the tags
In my PR last night I had one place that I missed updating unifi4 to unifi5.
I took the oppertunity to clean up the readme a bit while fixing it.

You'll also notice that I updated the tags referenced in the README with
what I thought made sense but let me know if you need me to update this PR
or feel free to pull this into your own branch and update this further.

I left the stable on unifi4 and oldstable on unifi3 as moving
that tag could suprise some users.

It seems as though the new builds on the docker hub don't have the expected tags
for some users.  As the unifi5 tag doesn't exist.

I wasn't sure how the automated build stuff worked so I left the README.md's in
each sub folder.  But they seem redundant and all but the main one could be
removed.