Support custom controller UI/API port for HEALTHCHECK
The HEALTHCHECK command hard-codes the default controller UI/API port.
This fails (shows the container status as unhealthy) if the user has
specified a custom port for it in the system properties file.

Make this work correctly in such cases by parsing the port from the file
if it is present and uncommented. In order to avoid a stupidly
complicated one-liner move the healthcheck command into a separate shell
script.