Add `instPkg` function for scripting.
`instPkg` will check for a package installed in dpkg and if it isn't installed yet use apt-get to install it.
Right now it takes a list of parameters and checks/installs each package individually. This behavior may change
in the future to allow other options like install a package from a local .deb instead of from the apt repositories.