Ensure ${DATADIR} and ${LOGDIR} exist on startup
This will attempt to create those directories and if running as a non root user set the permissions.
If for some reason a file has the name of one of those directories it will error. However you can set
the undocumented docker environment variable UNSAFE_IO to true and it will attempt to rm -rf the file
creating the directory. This is left off by default because... well it seems a bad idea to potentially rm
someones data directory without any input from them.