Write properties file after creating ${DATADIR}
Create the `system.properties` file after creating the missing
`${DATADIR}` directory. This file will only be created now if the
argument to the script is `unifi`, but that seems appropriate as the
properties file is only needed when starting the controller.

The first run of `docker-entrypoint.sh` on an empty unifi volume will
fail to create the `system.properties` file as the `data` directory
(`${DATADIR}`) on that volume is not created until after it tries to
write the `system.properties` file.

Issue: https://github.com/jacobalberty/unifi-docker/issues/401